Add a methos to BookList that adds new books, move addBtn function to BookList class as a method
